Solved

+++ Codeforces/oddDivisor.cpp
@@ -1,11 +1,11 @@
-// Problem Link: 
+// Problem Link: https://codeforces.com/contest/1475/problem/A
 
 #include<bits/stdc++.h>
 using namespace std;
 
 void oddDivisor()
 {
-    int n; 
+    long long int n; 
     cin>>n;
     if(n&(n-1)) cout<<"YES"<<endl;
     else cout<<"NO"<<endl;
